+++ system_modes/src/system_modes/mode_impl.cpp
@@ -14,18 +14,26 @@
 // limitations under the License.
 #include "system_modes/mode_impl.hpp"
 
+#include <map>
 #include <string>
 #include <vector>
 #include <utility>
 #include <exception>
 
+#include <lifecycle_msgs/msg/state.hpp>
+#include <lifecycle_msgs/msg/transition.hpp>
+
+using std::map;
 using std::pair;
 using std::mutex;
 using std::string;
 using std::vector;
+using std::to_string;
 using std::lock_guard;
 using std::out_of_range;
 using rclcpp::Parameter;
+using lifecycle_msgs::msg::State;
+using lifecycle_msgs::msg::Transition;
 
 namespace system_modes
 {
@@ -111,9 +119,9 @@ ModeImpl::add_parameters(const vector<Parameter> & parameters)
 void
 ModeImpl::set_parameter(const Parameter & parameter)
 {
-  std::string param_name = parameter.get_name();
+  string param_name = parameter.get_name();
   std::size_t foundr = parameter.get_name().rfind("ros__parameters");
-  if (foundr != std::string::npos) {
+  if (foundr != string::npos) {
     param_name = parameter.get_name().substr(foundr + strlen("ros__parameters") + 1);
   }
 
@@ -136,7 +144,7 @@ ModeImpl::set_parameters(const vector<Parameter> & parameters)
 
 void
 ModeImpl::add_part_mode(
-  const std::string & part,
+  const string & part,
   const StateAndMode stateAndMode)
 {
   this->part_modes_[part] = StateAndMode(stateAndMode.state, stateAndMode.mode);
@@ -144,7 +152,7 @@ ModeImpl::add_part_mode(
 
 void
 ModeImpl::set_part_mode(
-  const std::string & part,
+  const string & part,
   const StateAndMode stateAndMode)
 {
   if (this->part_modes_.find(part) == this->part_modes_.end()) {
@@ -156,7 +164,7 @@ ModeImpl::set_part_mode(
   this->add_part_mode(part, stateAndMode);
 }
 
-const std::vector<std::string>
+const vector<string>
 ModeImpl::get_parts() const
 {
   vector<string> results;
@@ -167,14 +175,102 @@ ModeImpl::get_parts() const
 }
 
 const StateAndMode
-ModeImpl::get_part_mode(const std::string & part) const
+ModeImpl::get_part_mode(const string & part) const
 {
   if (this->part_modes_.count(part)) {
     return this->part_modes_.at(part);
   } else {
-    throw std::out_of_range(
+    throw out_of_range(
             "Can't receive modes for part '" + part +
             "', part not specified.");
+  }
+}
+
+// TODO(anordman): Can we get this from the rcl default state machine?
+static const map<unsigned int, string> STATES_ = {
+  {State::PRIMARY_STATE_UNKNOWN, "unknown"},
+  {State::PRIMARY_STATE_UNCONFIGURED, "unconfigured"},
+  {State::PRIMARY_STATE_INACTIVE, "inactive"},
+  {State::PRIMARY_STATE_ACTIVE, "active"},
+  {State::PRIMARY_STATE_FINALIZED, "finalized"},
+  {State::TRANSITION_STATE_CONFIGURING, "configuring"},
+  {State::TRANSITION_STATE_CLEANINGUP, "cleaningup"},
+  {State::TRANSITION_STATE_SHUTTINGDOWN, "shuttingdown"},
+  {State::TRANSITION_STATE_ACTIVATING, "activating"},
+  {State::TRANSITION_STATE_DEACTIVATING, "deactivating"},
+  {State::TRANSITION_STATE_ERRORPROCESSING, "errorprocessing"}
+};
+
+static const map<unsigned int, string> TRANSITIONS_ = {
+  {Transition::TRANSITION_CREATE, "create"},
+  {Transition::TRANSITION_CONFIGURE, "configure"},
+  {Transition::TRANSITION_CLEANUP, "cleanup"},
+  {Transition::TRANSITION_ACTIVATE, "activate"},
+  {Transition::TRANSITION_DEACTIVATE, "deactivate"},
+  {Transition::TRANSITION_INACTIVE_SHUTDOWN, "inactive_shutdown"},
+  {Transition::TRANSITION_ACTIVE_SHUTDOWN, "active_shutdown"},
+  {Transition::TRANSITION_DESTROY, "destroy"}
+};
+
+static const map<unsigned int, unsigned int> GOAL_STATES_ = {
+  {Transition::TRANSITION_CREATE, State::PRIMARY_STATE_UNCONFIGURED},
+  {Transition::TRANSITION_CONFIGURE, State::PRIMARY_STATE_INACTIVE},
+  {Transition::TRANSITION_CLEANUP, State::PRIMARY_STATE_UNCONFIGURED},
+  {Transition::TRANSITION_ACTIVATE, State::PRIMARY_STATE_ACTIVE},
+  {Transition::TRANSITION_DEACTIVATE, State::PRIMARY_STATE_INACTIVE},
+  {Transition::TRANSITION_INACTIVE_SHUTDOWN, State::PRIMARY_STATE_FINALIZED},
+  {Transition::TRANSITION_ACTIVE_SHUTDOWN, State::PRIMARY_STATE_FINALIZED}
+};
+
+const string
+state_label_(unsigned int state_id)
+{
+  try {
+    return STATES_.at(state_id);
+  } catch (...) {
+    return "unknown";
+  }
+}
+
+unsigned int
+state_id_(const string & state_label)
+{
+  for (auto id : STATES_) {
+    if (id.second.compare(state_label) == 0) {
+      return id.first;
+    }
+  }
+  return 0;
+}
+
+const string
+transition_label_(unsigned int transition_id)
+{
+  try {
+    return TRANSITIONS_.at(transition_id);
+  } catch (...) {
+    throw out_of_range(string("Unknown transition id ") + to_string(transition_id));
+  }
+}
+
+unsigned int
+transition_id_(const string & transition_label)
+{
+  for (auto id : TRANSITIONS_) {
+    if (id.second.compare(transition_label) == 0) {
+      return id.first;
+    }
+  }
+  throw out_of_range("Unknown transition " + transition_label);
+}
+
+unsigned int
+goal_state_(unsigned int transition_id)
+{
+  try {
+    return GOAL_STATES_.at(transition_id);
+  } catch (...) {
+    throw out_of_range(string("Unknown transition id ") + to_string(transition_id));
   }
 }