+++ system_modes_examples/src/drive_base.cpp
@@ -18,11 +18,8 @@
 #include <lifecycle_msgs/msg/state.hpp>
 #include <lifecycle_msgs/msg/transition.hpp>
 
-#include <chrono>
-#include <iostream>
 #include <memory>
-#include <string>
-#include <thread>
+#include <vector>
 
 using lifecycle_msgs::msg::State;
 using lifecycle_msgs::msg::Transition;
@@ -39,33 +36,34 @@ namespace examples
 class DriveBase : public LifecycleNode
 {
 public:
-  explicit DriveBase()
+  DriveBase()
   : LifecycleNode("drive_base")
   {
     RCLCPP_INFO(get_logger(), "Constructed lifecycle node '%s'", this->get_name());
-    
+
     // Parameter declaration
-    this->declare_parameter("max_speed",
+    this->declare_parameter(
+      "max_speed",
       rclcpp::ParameterValue(rclcpp::PARAMETER_NOT_SET),
       rcl_interfaces::msg::ParameterDescriptor());
-    this->declare_parameter("controller",
+    this->declare_parameter(
+      "controller",
       rclcpp::ParameterValue(rclcpp::PARAMETER_NOT_SET),
       rcl_interfaces::msg::ParameterDescriptor());
 
     auto param_change_callback =
-    [this](std::vector<rclcpp::Parameter> parameters) -> rcl_interfaces::msg::SetParametersResult
-    {
-      auto result = rcl_interfaces::msg::SetParametersResult();
-      result.successful = true;
-      for (auto parameter : parameters) {
-        RCLCPP_INFO(this->get_logger(),
-          "parameter '%s' is now: %s",
-          parameter.get_name().c_str(),
-          parameter.value_to_string().c_str()
-          );
-      }
-      return result;
-    };
+      [this](std::vector<rclcpp::Parameter> parameters) -> rcl_interfaces::msg::SetParametersResult
+      {
+        auto result = rcl_interfaces::msg::SetParametersResult();
+        result.successful = true;
+        for (auto parameter : parameters) {
+          RCLCPP_INFO(this->get_logger(),
+            "parameter '%s' is now: %s",
+            parameter.get_name().c_str(),
+            parameter.value_to_string().c_str());
+        }
+        return result;
+      };
     this->set_on_parameters_set_callback(param_change_callback);
   }
 
@@ -78,7 +76,7 @@ public:
     RCLCPP_INFO(get_logger(), "on_configure()", this->get_name());
 
     return rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn::SUCCESS;
-  };
+  }
 
   rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn
   on_activate(const rclcpp_lifecycle::State &)
@@ -86,7 +84,7 @@ public:
     RCLCPP_INFO(get_logger(), "on_activate()", this->get_name());
 
     return rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn::SUCCESS;
-  };
+  }
 
   rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn
   on_deactivate(const rclcpp_lifecycle::State &)
@@ -94,7 +92,7 @@ public:
     RCLCPP_INFO(get_logger(), "on_deactivate()", this->get_name());
 
     return rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn::SUCCESS;
-  };
+  }
 
   rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn
   on_cleanup(const rclcpp_lifecycle::State &)
@@ -102,7 +100,7 @@ public:
     RCLCPP_INFO(get_logger(), "on_cleanup()", this->get_name());
 
     return rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn::SUCCESS;
-  };
+  }
 };
 
 }  // namespace examples